bug fix in Append() for nullable columns
If Append() of the nested column fails, the nulls column will be incremented despite nothing being appended.

It will also be marked as not null, causing an error to be thrown if this new "fake" value in the column is accessed later.

this fixes that by moving the appending of the nulls markers to the end of the Append() operation of the nested column